Moved remaining dev files
