--- dev/_scripts/CMakeLists.txt
@@ -1,7 +1,0 @@
-# TODO: Integrate with CDash https://github.com/jeffbaumes/cdash-examples
-# http://www.cmake.org/pipermail/cmake/2008-May/021864.html
-
-# find_package(PythonInterp REQUIRED) add_test(NAME some_test COMMAND
-# ${PYTHON_EXECUTABLE} ${CMAKE_CURRENT_BINARY_DIR}/run_test.py some_test)
-
-# CURRENTLY THIS DOES NOT DO ANYTHING
