--- dev/_scripts/purge_repos.sh
@@ -1,7 +1,0 @@
-cd ~/code
-rm -rf vtool
-rm -rf utool
-rm -rf plottool
-rm -rf guitool
-rm -rf hesaff
-rm -rf pyrf
