--- dev/optional-requirements.txt
@@ -1,35 +1,0 @@
-
-autopep8
-coverage>=3.7.1
-flake8
-
-GitPython>=2.0.8
-
-h5py
-line-profiler>=1.0
-memory_profiler
-
-mock>=1.3.0
-
-networkx>=1.10
-nose>=1.3.7
-pep8
-
-pyfiglet>=0.7.4
-pygco
-
-pyinstaller
-pylint>=1.5.4
-pyperclip>=1.5.7
-pytest>=2.9.2
-
-python-Levenshtein>=0.12.0
-quantumrandom>=1.9.0
--e git+https://github.com/tpircher/quine-mccluskey@master#egg=quine-mccluskey
-
-redbaron>=0.6.1
-selenium>=2.53.5
-setproctitle>=1.1.8
-Shapely>=1.5.7
-sympy
-wget
